+++ irc-connect/CHANGELOG.md
@@ -1,5 +1,9 @@
 # changelog
 
+## unreleased
+- `Stream::new_tcp` now requires a `SocketAddr` instead of a
+  `&SocketAddr`, since it is `Copy`
+
 ## 0.1.1 - 2025-07-28
 - fix documentation for socks
 - mark current socks behavior as deprecated