+++ irc-connect/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 - `Stream::new_tcp` now requires a `SocketAddr` instead of a
   `&SocketAddr`, since it is `Copy`
+- socks behavior switched around and now requires a `SocketAddr`
 
 ## 0.1.1 - 2025-07-28
 - fix documentation for socks