+++ irc-connect/CHANGELOG.md
@@ -4,6 +4,7 @@
 - `Stream::new_tcp` now requires a `SocketAddr` instead of a
   `&SocketAddr`, since it is `Copy`
 - socks behavior switched around and now requires a `SocketAddr`
+- `StreamBuilder` has been marked as `must_use`
 
 ## 0.1.1 - 2025-07-28
 - fix documentation for socks